switches js.Eval to use output.eval_js and message to use colab.google._message
Incidentally this fixes the bug in this CL, because webcomponent_helper was expecting message.MessageError exception but was getting _message.MessageError instead.

Switches colab_message_id to be integer (rather than uuid) to keep compatibility with internal tests. Alternatively we could modify fix our internal tests (which would be slightly more work), but it is not clear what the benefits are.

RELNOTES: n/a
PiperOrigin-RevId: 176741532