External: auth: Work-around for the authorization prompt race described in b/70391156#comment9.
The main idea is to bundle the link with the getpass prompt. For this to work, we need to
linkify when constructing the prompt. (Right now, if you do a getpass.getpass('http://google.com'),
it will linkify, but only after the prompt is completed.)

The fixes are ~5 LOC. The larger part of this change is for testing:
datalab/integration:runlocal now supports running a notebook probe
(here, auth.ipynb) against a local serving stack. This is necessary
to test notebook probes that depend on modifications to both the
FE and BE container.

Rather than building a new container, I tested by running ./update_container.sh
against the existing staging container in the interval between :runlocal
starting the container and when the probe began.

Testing video:
https://drive.google.com/open?id=1ZaKWb_bhwFIChmFGLYCPbhUpEHMqU6pT

While inspecting logs, I noticed some analytics requests sneaking through
via HTTPS. Those have a :443 host suffix and are now filtered as well.
(I did not see any doubleclick requests, however, as suggested by b/65599348.)

RELNOTES: n/a
PiperOrigin-RevId: 179890529