Use existing credentials in external colab if they exist.
Currently, we run the oauth2 dance for a user per *process*, whereas we really
only need to do it per *VM*. No one likes doing the oauth2 dance
unnecessarily.

The fix in this CL is to set the GOOGLE_APPLICATION_CREDENTIALS early, which
then skips the dance if a token is already in place. We leave an initial check
in place for testing purposes.

RELNOTES: n/a
PiperOrigin-RevId: 180628256