Switch the name of our colabtools build directory.
It turns out that argo has (wisely) decided to respect `.gitignore` when
considering which files to package up for use in cloud builds. Unfortunately
for us, we chose to keep our configuration in a directory called `build`,
which is ignored by our `.gitignore`.

This CL cleans things up by moving `build` to `cloudbuild` and updating
`build.sh` accordingly. This isn't just to sate argo: we ignore `build`
because Python stages files there when building, and we don't want to mix
those with files we use.

(This is split off a larger change to build our containers with argo.)

Tested:
  http://pantheon/gcr/builds/bed65e08-a454-460e-8a01-78d1b61ee96e?project=colab-datalab&organizationId=433637338589

RELNOTES: n/a
PiperOrigin-RevId: 184149309