Add a missing > in some inline JS, and fix some formatting.
This was reported by an external user; I am amazed Chrome was so forgiving in
the first place. In a perfect world, this will help with some of the upload
flakiness we've seen reported.

While I was here, I also moved the `%` symbol in the status message (it was
displaying `%100` instead of `100%`).

This picked up a new version of `pandas-gbq`, which in turn requires a newer
version of `google-cloud-bigquery`
(https://github.com/pydata/pandas-gbq/commit/183daf1efc7d4dd57ac266825d13ea9d6dd86cb6).
I think that's actually a fishy thing to do in a bugfix release, but onward
and upward, I guess.

Pushed docker image:
  gcr.io/colab-datalab/datalab:auto-186773299-2018-02-23T10-15-41-965571869-08-00

RELNOTES: n/a
PiperOrigin-RevId: 186800691