Allow kernel to provide structured info replies and metadata for completions.
For inspect_requests, the response was a ansi-encoded display-formatted string which dropped a lot of the metadata included in the v2 kernel. This was explained in https://github.com/ipython/ipython/wiki/IPEP-24:-completion-and-object_info.

This CL just re-adds the structured response along-side the display-formatted string.

In addition, it allows requesting additional metadata for the completion requests, similar to cl/193991146.

RELNOTES: N/A
PiperOrigin-RevId: 196269244