Automated g4 rollback of changelist 218574913.
*** Reason for rollback ***

This CL was woefully misguided.

There were no warnings in play, just logging at loglevel warning. /facepalm

*** Original change description ***

Silence warnings from google.auth about missing project ID.

google.auth is primarily interested in the case of GCP, where it always makes
sense to have a cloud project in play. To that end, if the ADC include a
credential but no project, google.auth logs a warning that the user probably
wants to set a project.

However, for colab, this isn't as clear-cut: we often want to use credentials
for non-cloud APIs, where projects don't matter. Rather than show this warning
to the user, we instead squel...

***

PiperOrigin-RevId: 218685214