Adjust `PATH` and `PYTHONPATH` from `%tensorflow_version`
This is needed for the `%tensorboard` magic to work properly on
TensorFlow 2.x: the version of TensorBoard launched as a subprocess must
be consistent with the version of the `tensorboard` Python module.

Tested:
Make a new notebook. Paste the contents of `_tensorflow_magics.py` into
a cell, and execute it. In a new cell, also execute

```
import IPython
_register_magics(IPython.get_ipython())
```

Then, run:

```
!tensorboard --version
%tensorflow_version 2.x
!tensorboard --version
%tensorflow_version 1.x
!tensorboard --version

%tensorflow_version 2.x
%load_ext tensorboard
%tensorboard --logdir /tmp/whatever
```

Observe that the output includes

```
1.14.0
TensorFlow 2.x selected.
1.14.0a20190603
TensorFlow 1.x selected.
1.14.0
```

and then a working TensorBoard iframe.

To see the behavior prior to this patch, reset all runtimes and execute
the last cell only; observe that the version printed is `1.14.0` in all
cases (which is incorrect), and that the TensorBoard iframe is replaced
by a bunch of log spam about a version mismatch.

PiperOrigin-RevId: 260171791